Convert cypress tests to test-it tests
