--- cypress/integration/2d/ogmo/tiles/test.js
@@ -1,19 +1,0 @@
-import { createRoot, useNewComponent, Canvas, Ogmo } from "@hex-engine/2d";
-import project from "./project.ogmo";
-import level from "./level.json";
-
-it("renders correctly", () => {
-  cy.createCanvas().then(($canvas) => {
-    createRoot(() => {
-      const canvas = useNewComponent(() =>
-        Canvas({ backgroundColor: "white", element: $canvas[0] })
-      );
-      canvas.fullscreen();
-
-      const ogmo = useNewComponent(() => Ogmo.Project(project, {}));
-
-      ogmo.useLevel(level);
-    });
-  });
-  cy.matchImageSnapshot();
-});
