+++ packages/integration-tests/src/ogmo/decals/level.json
