+++ packages/integration-tests/src/ogmo/decals/project.ogmo
