+++ packages/integration-tests/src/ogmo/decals/tiles.png
