+++ packages/integration-tests/src/ogmo/entities/tiles.png
