+++ packages/integration-tests/src/ogmo/layer ordering/tiles.png
