+++ packages/integration-tests/src/ogmo/tiles/tiles.png
