+++ packages/scripts/package.json
@@ -29,8 +29,8 @@
     "html-webpack-plugin": "^3.2.0",
     "mime-types": "^2.1.26",
     "on-exit": "^1.0.1",
-    "ogmo-level-json-loader": "^0.1.0",
-    "ogmo-project-loader": "^0.1.0",
+    "ogmo-level-json-loader": "^0.2.0",
+    "ogmo-project-loader": "^0.2.0",
     "react-dev-utils": "^10.0.0",
     "regenerator-runtime": "^0.13.3",
     "resolve": "^1.14.2",
